files replaced

+++ src/features/cart/CartSlice.js
@@ -2,6 +2,16 @@ import { createSlice } from '@reduxjs/toolkit';
 
 const initialState = {
   cart: [],
+
+  // cart: [
+  //   {
+  //     pizzaId: 12,
+  //     name: 'Mediterranean',
+  //     quantity: 2,
+  //     unitPrice: 16,
+  //     totalPrice: 32,
+  //   },
+  // ],
 };
 
 const cartSlice = createSlice({
@@ -9,24 +19,28 @@ const cartSlice = createSlice({
   initialState,
   reducers: {
     addItem(state, action) {
+      // payload = newItem
       state.cart.push(action.payload);
     },
     deleteItem(state, action) {
+      // payload = pizzaId
       state.cart = state.cart.filter((item) => item.pizzaId !== action.payload);
     },
     increaseItemQuantity(state, action) {
+      // payload = pizzaId
       const item = state.cart.find((item) => item.pizzaId === action.payload);
 
       item.quantity++;
       item.totalPrice = item.quantity * item.unitPrice;
     },
     decreaseItemQuantity(state, action) {
+      // payload = pizzaId
       const item = state.cart.find((item) => item.pizzaId === action.payload);
+
       item.quantity--;
       item.totalPrice = item.quantity * item.unitPrice;
 
-      if (item.quantity === 0)
-        cartSlice.caseReducers.deleteItem(state, action)
+      if (item.quantity === 0) cartSlice.caseReducers.deleteItem(state, action);
     },
     clearCart(state) {
       state.cart = [];
@@ -44,8 +58,6 @@ export const {
 
 export default cartSlice.reducer;
 
-export const getUsername = (state) => state.user.username;
-
 export const getCart = (state) => state.cart.cart;
 
 export const getTotalCartQuantity = (state) =>
@@ -56,3 +68,5 @@ export const getTotalCartPrice = (state) =>
 
 export const getCurrentQuantityById = (id) => (state) =>
   state.cart.cart.find((item) => item.pizzaId === id)?.quantity ?? 0;
+
+// 'reselect'
